refactoring: renamed DeepFakeBot cog to CoreCommands, preparing to add cooldowns, moved app declaration into run_app() (allows database connection messages to reach the logger)
